requirements.txt
git+https://github.com/bytedance/MVDream
accelerate==0.23.0
diffusers==0.25.0
peft==0.1.0
wandb
siren-pytorch
jaxtyping
typeguard
opencv-python
scikit-learn
jupyterlab
matplotlib
scipy
pandas
datasets
huggingface_hub==0.25.0
git+https://github.com/facebookresearch/pytorch3d.git@stable
torchpq
faiss-gpu
cupy-cuda12x
transformers==4.42.0
controlnet_aux
onnxruntime-gpu

# for threestudio
lightning==2.0.0
omegaconf==2.3.0
git+https://github.com/KAIR-BAIR/nerfacc.git@v0.5.2
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
envlight
libigl
tensorboard
imageio>=2.28.0
imageio[ffmpeg]
git+https://github.com/NVlabs/nvdiffrast.git
xatlas
trimesh[easy]
networkx
pysdf
PyMCubes
gradio==4.11.0
torchmetrics
xformers
bitsandbytes
sentencepiece
safetensors
einops
kornia
taming-transformers-rom1504
git+https://github.com/openai/CLIP.git